Adds failing testcase to clarify expected functionality of the decodeHex method
In our project we wanted to identify if a given string was a valid
HashId. The first version of this utilised `decodeHex` and then tested
to see if the resulting string was empty or not. This validation did
not work for values under 16 thus leading me in the direction of this
strange behaviour of `decodeHex`.

As far as I can tell this has been this way since it was introduced in
43b258a7b9121046b18f2fe27fbfddd12bbb43cb what is it's intended usage?