Added "separators" customization, NULL pre-checks, and Javadoc comment blocks to all constructors
Added a new Hashids constructor to support customized "separators" plus a pre-check for NULL that prevents a NullException by defaulting to DEFAULT_SEPS (this behaviour is modeled after the "salt" pre-check for NULL).

Added a pre-check for NULL for the custom "alphabet" that prevents a NullException by defaulting to DEFAULT_ALPHABET (this behaviour is also modeled after the "salt" pre-check for NULL).

Added rudimentary comments for existing constructors for better Javadoc output.  Added relevant details about the new constructor that supports custom "separators" (should this also be mentioned in the documentation on the web site?).